ACT-3: fix calculation to match oasys (#10)

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/arnsriskactuarialapi/dto/OGRS3Object.kt
@@ -2,8 +2,8 @@ package uk.gov.justice.digital.hmpps.arnsriskactuarialapi.dto
 
 data class OGRS3Object(
   val algorithmVersion: String,
-  val ogrs3OneYear: Double?,
-  val ogrs3TwoYear: Double?,
+  val ogrs3OneYear: Int?,
+  val ogrs3TwoYear: Int?,
   val band: RiskBand?,
   val validationError: List<ValidationErrorResponse>?,
 )
